fixed navbar for mobile hopefully.... seems the chrome devtools do not perfectly simulate mobile devices as my physical device has a bug that is not showing in testing environments.
So... further polished navbar for mobile.
Planning to change up project display to include demo button and programming language names (as icon is not enough for the general audience). Also need to change the timeline thing.... i had an idea of what to put there but can't think of it now.

So update projects tiles
Progress meter thing
Change timeline... maybe just put a scroller of something else